Unit tests back to snuff

+++ cqrs/cqrs.go
@@ -8,13 +8,13 @@ import (
 
 const MESSAGE_TYPE_MASK = 0x80000000
 
-func C(domain uint32, version uint32, typeId uint32) uint64 {
-	return (uint64(domain) << 32) | uint64(MESSAGE_TYPE_MASK) | uint64(version&0x7FFF<<16) | uint64(typeId&0xFFFF)
+func C(domain uint32, version uint64, typeId uint64) uint64 {
+	return (uint64(domain) << 32) | uint64(MESSAGE_TYPE_MASK) | (version & 0x7FFF << 16) | (typeId & 0xFFFF)
 
 }
 
-func E(domain uint32, version uint32, typeId uint32) uint64 {
-	return (uint64(domain) << 32) | uint64(version&0x7FFF<<16) | uint64(typeId&0xFFFF)
+func E(domain uint32, version uint64, typeId uint64) uint64 {
+	return (uint64(domain) << 32) | (version & 0x7FFF << 16) | (typeId & 0xFFFF)
 }
 
 type AggregateLoader interface {
